fix: address 4 critical code review issues
- Use BigDecimal for all monetary values (commission, totalRepayable,
  loanAmount, riskBand multipliers) to prevent floating-point rounding
  errors in financial calculations. QuoteResponse serializes with
  @JsonFormat(shape=STRING) to preserve 2 decimal places in JSON.

- Remove hardcoded spring.profiles.active: dev from base application.yml.
  Deployments no longer silently wire MockCommissionApiAdapter; the dev
  profile must be activated explicitly via SPRING_PROFILES_ACTIVE=dev.

- Move UpstreamApiException from adapter.outbound.mock to
  adapter.outbound.http (CommissionApiExceptions.kt) so production code
  no longer has a package-dependency inversion on a dev/mock class.

- Remove redundant .timeout(Duration.ofSeconds(10)) operator from
  CommissionApiAdapter reactive chain. The WebClient-level responseTimeout
  (Netty ReadTimeoutException) is the single timeout mechanism, correctly
  mapped to CommissionApiTimeoutException -> 504. The redundant operator
  was causing 500 responses on timeout instead of 504.

All tests updated and passing: 43 backend, 53 frontend, 7 E2E.

+++ server/src/main/kotlin/com/example/commissionquote/domain/model/QuoteResult.kt
@@ -1,7 +1,9 @@
 package com.example.commissionquote.domain.model
 
+import java.math.BigDecimal
+
 data class QuoteResult(
     val quoteId: String,
-    val commission: Double,
-    val totalRepayable: Double,
+    val commission: BigDecimal,
+    val totalRepayable: BigDecimal,
 )
